requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dlib==19.24.0
executing==1.1.0
Flask==2.2.2
fonttools==4.37.3
idna==3.4
imageio==2.22.0
imgaug==0.4.0
importlib-metadata==4.12.0
ipyplot==1.1.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.30
networkx==2.8.6
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-multiflow==0.5.3
scikit-multilearn==0.2.0
scipy==1.9.1
Shapely==1.8.4
shortuuid==1.0.9
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.5.1
threadpoolctl==3.1.0
tifffile==2022.8.12
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
